# openpredict/openapi.yml
openapi: 3.0.1
info:
  description:
    "Get predicted targets for a given entity: the **potential drugs treating a given disease**, or the **potential diseases a given drug could treat**.
    \n\n* Use the `/predict` operation to easily retrieve predictions for a given entity (operation annotated for the [BioThings Explorer](https://x-bte-extension.readthedocs.io/en/latest/x-bte-kgs-operations.html)).
    \n\n* Predictions are currently produced using the [PREDICT method](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3159979/) from the [fair-workflows/openpredict](https://github.com/fair-workflows/openpredict) project.
    \n\n* You can also install the [OpenPredict Python package](https://pypi.org/project/openpredict/) to train and serve a new model yourself.
    \n\n* More documentation available at [github.com/MaastrichtU-IDS/translator-openpredict](https://github.com/MaastrichtU-IDS/translator-openpredict)
    \n\n[![Test production API](https://github.com/MaastrichtU-IDS/translator-openpredict/actions/workflows/run-tests-prod.yml/badge.svg)](https://github.com/MaastrichtU-IDS/translator-openpredict/actions/workflows/run-tests-prod.yml)
    \n\nService supported by the [NCATS Translator project](https://ncats.nih.gov/translator/about)"
  version: 1.0.0
  title: OpenPredict API
  contact:
    email: vincent.emonet@maastrichtuniversity.nl
  license:
    name: MIT license
    url: https://opensource.org/licenses/MIT
...
  x-translator:
    component: KP
    team:
...
  x-trapi:
...
servers:
...
paths:
  /features:
    get:
      operationId: openpredict.openpredict_api.get_features
      summary: Return the features trained in the models
      description: "Return the features trained in the model, for Drugs, Diseases or Both "
      tags:
        - openpredict
      parameters:
        - name: type
          in: query
          description: Type of the entities in the features
          required: true
          schema:
            type: string
            enum: ["All", "Drugs", "Diseases", "Both"]
            default: "All"
      responses:
        "200":
          description: OK
          content:
            application/json: {}
...
  /embedding:
    post:
      operationId: openpredict.openpredict_api.post_embedding
      summary: Upload your embedding for drugs or diseases
      description: |-
        Upload your embedding file:  select which types do you have in the embeddings, Drugs, Diseases or Both. 
        1. provided embeddings will be added to the model
        2. the model will be retrained
        3. the model evaluation will be stored in a triplestore
      tags:
        - openpredict
      responses:
        "200":
          description: OK
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                embedding_file:
                  description: The embedding file to be uploaded
                  type: string
                  format: binary
      parameters:
        - name: apikey
          in: query
          description: API key
          required: true
          schema:
            type: string
        - name: types
          in: query
          description: Types of the entities in the embedding
          required: true
          schema:
            type: string
            enum: ["Drugs", "Diseases", "Both"]
            default: "Both"
        - name: emb_name
          in: query
          description: Name of embedding (eg. TransE_DrugBank_Drug)
          required: true
          schema:
            type: string
        - name: description
          in: query
          description: Describe how the embeddings was obtained (eg. rdf2vec on drugbank/omim data)
          required: true
          schema:
            type: string
        - name: model_id
          in: query
          description: Provide the ID of the run output you want to use as base model to add the embeddings
          required: true
          example: openpredict-baseline-omim-drugbank
          schema:
            type: string
        # - name: apikey
        #   in: query
        #   description: APIKEY required to upload embeddings
        #   required: true
        #   schema:
        #     type: string
  /similarity:
    get:
      operationId: openpredict.openpredict_api.get_similarity
      summary: Return similar entities for a given drug or disease
      description: |-
        select which type entities do you expect in the result, Drugs, Diseases
        1. provide a drug or a disease
        2. it will return most similar entities based on similarity model
      tags:
        - openpredict
      responses:
        "200":
          description: OK
      parameters:
        - name: types
          in: query
          description: Types of the entities you want to return
          required: true
          schema:
            type: string
            enum: ["Drugs", "Diseases"]
            default: "Drugs"
        - name: drug_id
          in: query
          description: CURIE of the query drug to get similar entities (e.g. DRUGBANK:DB00394)
          example: DRUGBANK:DB00394
          required: false
          schema:
            type: string
        - name: disease_id
          in: query
          description: CURIE of the query disease to get similar entities (e.g. OMIM:246300). Ignored if drug_id provided.
          # example: OMIM:246300
          required: false
          schema:
            type: string
        - name: model_id
          in: query
          description: Provide the ID of the embedding model you want to use as base model to calculate the similarities
          required: true
          example: drugs_fp_embed.txt
          schema:
            type: string
        - name: min_score
          in: query
          description: Minimum score of predictions to return (0 to 1)
          required: false
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 1
        - name: max_score
          in: query
          description: Maximum score of predictions to return (0 to 1)
          required: false
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 1
        - name: n_results
          in: query
          description: Maximum number of results to return, ordered from best score (e.g. 42)
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
  /predict:
    get:
      operationId: openpredict.openpredict_api.get_predict
      summary: Get predicted targets for a given entity
      description:
        "Return the predicted targets for a given entity: drug (DrugBank ID) or disease (OMIM ID), with confidence scores.
        Only a drug_id or a disease_id can be provided, the disease_id will be ignored if drug_id is provided
        This operation is annotated with x-bte-kgs-operations, and follow the BioThings API recommendations."
      tags:
        - biothings
      parameters:
        - name: drug_id
          in: query
          description: CURIE of the drug to get prediction of diseases it could treat (e.g. DRUGBANK:DB00394)
          example: DRUGBANK:DB00394
          required: false
          schema:
            type: string
        - name: disease_id
          in: query
          description: CURIE of the disease to get prediction of drugs it could be treated with (e.g. OMIM:246300). Ignored if drug_id provided.
          # example: OMIM:246300
          required: false
          schema:
            type: string
        - name: model_id
          in: query
          description: Provide the ID of the run output you want to use as model to use to predict associations
          required: false
          example: openpredict-baseline-omim-drugbank
          schema:
            type: string
            # default: "Predict OMIM-DrugBank"
            # enum: ["Predict OMIM-DrugBank"]
        - name: min_score
          in: query
          description: Minimum score of predictions to return (0 to 1)
          required: false
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 1
        - name: max_score
          in: query
          description: Maximum score of predictions to return (0 to 1)
          required: false
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 1
        - name: n_results
          in: query
          description: Maximum number of results to return, ordered from best score (e.g. 42)
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        "200":
          description: Successful Operation
          content:
            application/json: {}
            # application/xml: {}
            # application/ld+json: {}
            # text/csv: {}
            # text/tab-separated-values: {}
      x-bte-kgs-operations:
        - inputs:
            - id: biolink:DRUGBANK
              semantic: biolink:ChemicalSubstance
          outputs:
            - id: biolink:OMIM
              semantic: biolink:Disease
          parameters:
            drug_id: "{inputs[0]}"
          predicate: biolink:treats
          supportBatch: False
          responseMapping:
            OMIM: hits.id
            # NCBIGene: matchedTerms.interactions.geneEntrezId
            # publication: matchedTerms.interactions.pmids
        - inputs:
            - id: biolink:OMIM
              semantic: biolink:Disease
          outputs:
            - id: biolink:DRUGBANK
              semantic: biolink:ChemicalSubstance
          parameters:
            disease_id: "{inputs[0]}"
          predicate: biolink:treated_by
          supportBatch: False
          responseMapping:
            DRUGBANK: hits.id
            # CHEMBL.COMPOUND: matchedTerms.interactions.drugChemblId
            # publication: results.source.id
...
  /query:
    post:
      operationId: openpredict.openpredict_api.post_reasoner_predict
      tags:
        - query
      summary: Initiate a query and wait to receive a Response
      description: ""
      requestBody:
        description: Query information to be submitted
        required: true
        content:
          application/json:
            schema:
              # $ref: '#/definitions/RequestBody'
              $ref: "#/components/schemas/Query"
      responses:
        "200":
          description: >-
            OK. There may or may not be results. Note that some of the provided
            identifiers may not have been recognized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: >-
            Bad request. The request is invalid according to this OpenAPI
            schema OR a specific identifier is believed to be invalid somehow
            (not just unrecognized).
          content:
            application/json:
              schema:
                type: string
        "500":
          description: >-
            Internal server error.
          content:
            application/json:
              schema:
                type: string
        "501":
          description: >-
            Not implemented.
          content:
            application/json:
              schema:
                type: string
  # /asyncquery:
  #   post:
  #     operationId: openpredict.openpredict_api.async_reasoner_predict
  #     tags:
  #       - asyncquery
  #     summary: Initiate a query with a callback to receive the response
  #     description: ''
  #     requestBody:
  #       description: Query information to be submitted
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/AsyncQuery'
  #     responses:
  #       '200':
  #         description: >-
  #           The query is accepted for processing and the Response will be
  #           sent to the callback url when complete.
  #         content:
  #           application/json:
  #             schema:
  #               type: string
  #       '400':
  #         description: >-
  #           Bad request. The request is invalid according to this OpenAPI
  #           schema OR a specific identifier is believed to be invalid somehow
  #           (not just unrecognized).
  #         content:
  #           application/json:
  #             schema:
  #               type: string
  #       '500':
  #         description: >-
  #           Internal server error.
  #         content:
  #           application/json:
  #             schema:
  #               type: string
  #       '501':
  #         description: >-
  #           Not implemented.
  #         content:
  #           application/json:
  #             schema:
  #               type: string
components:
  schemas:
...
    Query:
      example:
        query_options:
          n_results: 10
          min_score: 0.5
          max_score: 1
        message:
          query_graph:
            nodes:
              n0:
                categories:
                  - "biolink:Disease"
                ids:
                  - OMIM:246300
                  - MONDO:0007190
              n1:
                categories:
                  - "biolink:Drug"
                # ids: DRUGBANK:DB00394
            edges:
              e01:
                subject: n0
                object: n1
                predicates:
                  - "biolink:treated_by"
      description: >-
        The Query class is used to package a user request for information. A
        Query object consists of a required Message object with optional
        additional properties. Additional properties are intended to convey
        implementation-specific or query-independent parameters. For example,
        an additional property specifying a log level could allow a user to
        override the default log level in order to receive more fine-grained
        log information when debugging an issue.
      x-body-name: request_body
      type: object
      properties:
        message:
          $ref: "#/components/schemas/Message"
          description: >-
            The query Message is a serialization of the user request. Content
            of the Message object depends on the intended TRAPI operation. For
            example, the fill operation requires a non-empty query_graph field
            as part of the Message, whereas other operations, e.g. overlay,
            require non-empty results and knowledge_graph fields.
        log_level:
          description: The least critical level of logs to return
          allOf:
            - $ref: "#/components/schemas/LogLevel"
          nullable: true
        # workflow:
        #   description: List of workflow steps to be executed.
        #   $ref: http://standards.ncats.io/workflow/1.0.0/schema
      additionalProperties: true
      required:
        - message
    AsyncQuery:
      description: >-
        The AsyncQuery class is effectively the same as the Query class but
        it requires a callback property.
      x-body-name: request_body
      type: object
      properties:
        callback:
          type: string
          format: uri
          pattern: ^https?://
          description: >-
            Upon completion, this server will send a POST request to the
            callback URL with `Content-Type: application/json` header and
            request body containing a JSON-encoded `Response` object.
            The server MAY POST `Response` objects before work is fully
            complete to provide interim results with a Response.status
            value of 'Running'. If a POST operation to the callback URL
            does not succeed, the server SHOULD retry the POST at least
            once.
        message:
          $ref: "#/components/schemas/Message"
          description: >-
            The query Message is a serialization of the user request. Content
            of the Message object depends on the intended TRAPI operation. For
            example, the fill operation requires a non-empty query_graph field
            as part of the Message, whereas other operations, e.g. overlay,
            require non-empty results and knowledge_graph fields.
        log_level:
          description: The least critical level of logs to return
          allOf:
            - $ref: "#/components/schemas/LogLevel"
          nullable: true
        # workflow:
        #   description: List of workflow steps to be executed.
        #   $ref: http://standards.ncats.io/workflow/1.0.0/schema
      additionalProperties: true
      required:
        - callback
        - message
    Response:
      type: object
      description: >-
        The Response object contains the main payload when a TRAPI query
        endpoint interprets and responds to the submitted query successfully
        (i.e., HTTP Status Code 200). The message property contains the
        knowledge of the response (query graph, knowledge graph, and results).
        The status, description, and logs properties provide additional details
        about the response.
      properties:
        message:
          description: >-
            Contains the knowledge of the response (query graph, knowledge
            graph, and results).
          $ref: "#/components/schemas/Message"
        status:
          description: >-
            One of a standardized set of short codes,
            e.g. Success, QueryNotTraversable, KPsNotAvailable
          type: string
          example: Success
          nullable: true
        description:
          description: A brief human-readable description of the outcome
          type: string
          example: Success. 42 results found.
          nullable: true
        logs:
          description: >-
            Log entries containing errors, warnings, debugging information, etc
          type: array
          items:
            $ref: "#/components/schemas/LogEntry"
          nullable: true
        # workflow:
        #   description: List of workflow steps that were executed.
        #   $ref: http://standards.ncats.io/workflow/1.0.0/schema
      additionalProperties: true
      required:
        - message
    Message:
      description: >-
        The message object holds the main content of a Query or a Response in
        three properties: query_graph, results, and knowledge_graph.
        The query_graph property contains the query configuration, the results
        property contains any answers that are returned by the service,
        and knowledge_graph property contains lists of edges and nodes in the
        thought graph corresponding to this message. The content of these
        properties is context-dependent to the encompassing object and
        the TRAPI operation requested.
      type: object
      properties:
        results:
          description: >-
            List of all returned Result objects for the query posed.
            The list SHOULD NOT be assumed to be ordered. The 'score' property,
             if present, MAY be used to infer result rankings.
          type: array
          items:
            $ref: "#/components/schemas/Result"
          nullable: true
        query_graph:
          description: >-
            QueryGraph object that contains a serialization of a query in the
            form of a graph
          allOf:
            - $ref: "#/components/schemas/QueryGraph"
          nullable: true
        knowledge_graph:
          description: >-
            KnowledgeGraph object that contains lists of nodes and edges
            in the thought graph corresponding to the message
          allOf:
            - $ref: "#/components/schemas/KnowledgeGraph"
          nullable: true
      additionalProperties: false
    LogEntry:
      description: >-
        The LogEntry object contains information useful for tracing
        and debugging across Translator components.  Although an
        individual component (for example, an ARA or KP) may have its
        own logging and debugging infrastructure, this internal
        information is not, in general, available to other components.
        In addition to a timestamp and logging level, LogEntry
        includes a string intended to be read by a human, along with
        one of a standardized set of codes describing the condition of
        the component sending the message.
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp in ISO 8601 format
          example: "2020-09-03T18:13:49+00:00"
          nullable: true
        level:
          allOf:
            - $ref: "#/components/schemas/LogLevel"
          nullable: true
        code:
          type: string
          description: >-
            One of a standardized set of short codes
            e.g. QueryNotTraversable, KPNotAvailable, KPResponseMalformed
          nullable: true
        message:
          type: string
          description: A human-readable log message
          nullable: true
      additionalProperties: true
    LogLevel:
      type: string
      description: Logging level
      enum:
        - ERROR
        - WARNING
        - INFO
        - DEBUG
    Result:
      type: object
      description: >-
        A Result object specifies the nodes and edges in the knowledge graph
        that satisfy the structure or conditions of a user-submitted query
        graph. It must contain a NodeBindings object (list of query graph node
        to knowledge graph node mappings) and an EdgeBindings object (list of
        query graph edge to knowledge graph edge mappings).
      properties:
        node_bindings:
          type: object
          description: >-
            The dictionary of Input Query Graph to Result Knowledge Graph node
            bindings where the dictionary keys are the key identifiers of the
            Query Graph nodes and the associated values of those keys are
            instances of NodeBinding schema type (see below). This value is an
            array of NodeBindings since a given query node may have multiple
            knowledge graph Node bindings in the result.
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/NodeBinding"
        edge_bindings:
          type: object
          description: >-
            The dictionary of Input Query Graph to Result Knowledge Graph edge
            bindings where the dictionary keys are the key identifiers of the
            Query Graph edges and the associated values of those keys are
            instances of EdgeBinding schema type (see below). This value is an
            array of EdgeBindings since a given query edge may resolve to
            multiple knowledge graph edges in the result.
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/EdgeBinding"
        score:
          type: number
          format: float
          example: 163.233
          description: >-
            A numerical score associated with this result indicating the
            relevance or confidence of this result relative to others in the
            returned set. Higher MUST be better.
      additionalProperties: true
      required:
        - node_bindings
        - edge_bindings
...