# openpredict/openpredict_api.py
import os
import connexion
import logging
from datetime import datetime
from openpredict.openpredict_utils import init_openpredict_dir
from openpredict.rdf_utils import init_triplestore, retrieve_features, retrieve_models
from openpredict.openpredict_model import addEmbedding, get_predictions, get_similarities
from openpredict.openpredict_model import load_similarity_embedding_models, load_treats_embedding_models, load_treats_classifier
from openpredict.reasonerapi_parser import typed_results_to_reasonerapi
from flask_cors import CORS
from flask_reverse_proxy_fix.middleware import ReverseProxyPrefixFix
import pkg_resources
# from gensim.models import KeyedVectors
# import asyncio
# import aiohttp
# from aiohttp import web

# all_emb_vectors = None
treats_features = None
similarity_features = None
treats_classifier = None

# self.treats_features = load_treats_embedding_models(treats_model)
# self.treats_classifier = load_treats_classifier(treats_model)
# self.similarity_features = load_similarity_embedding_models()

# TODO: check using FastAPI: https://github.com/NCATS-Tangerine/icees-api/blob/master/icees_api/trapi.py
def start_api(port=8808, debug=False, start_spark=True):
    """Start the Translator OpenPredict API using [zalando/connexion](https://github.com/zalando/connexion) and the `openapi.yml` definition

    :param port: Port of the OpenPredict API, defaults to 8808
    :param debug: Run in debug mode, defaults to False
    :param start_spark: Start a local Spark cluster, default to true
    """
    print("Starting the \033[1mTranslator OpenPredict API\033[0m 🔮🐍")

    init_openpredict_dir()
    init_triplestore()
    baseline_model = 'openpredict-baseline-omim-drugbank'
    global similarity_features
    similarity_features = load_similarity_embedding_models()
    global treats_features
    treats_features = load_treats_embedding_models(baseline_model)
    global treats_classifier
    treats_classifier = load_treats_classifier(baseline_model)

    if debug:
        # Run in development mode
        deployment_server = 'flask'
        logging.basicConfig(level=logging.DEBUG)
        print("Development deployment using \033[1mFlask\033[0m 🧪")
        print("Debug enabled 🐞 - The API will reload automatically at each change 🔃")
    else:
        # Run in productiom with tornado (also available: gevent)
        deployment_server = 'tornado'
        # logging.basicConfig(level=logging.INFO)
        logging.basicConfig(level=logging.DEBUG)
        print("Production deployment using \033[1mTornado\033[0m 🌪️")

    api = connexion.App(__name__, options={"swagger_url": ""})
    # Try Aio for sync, cf. https://github.com/zalando/connexion/issues/1277
    # api = connexion.AioHttpApp(__name__, options={"swagger_url": ""}, only_one_api=True)
    # api = connexion.App(__name__, options={"swagger_url": ""}, arguments={'server_url': server_url})

    api.add_api('openapi.yml')
    # api.add_api('openapi.yml', arguments={'server_url': server_url}, validate_responses=True, options={"disable_servers_overwrite": True})

    # Add CORS support
    CORS(api.app)

    # Fix to avoid empty list of servers for nginx-proxy deployments
    if os.getenv('LETSENCRYPT_HOST'):
        server_url = 'https://' + os.getenv('LETSENCRYPT_HOST')
        api.app.config['REVERSE_PROXY_PATH'] = server_url
        # api.app.config['REVERSE_PROXY_PATH'] = '/api'
        ReverseProxyPrefixFix(api.app)
    elif os.getenv('VIRTUAL_HOST'):
        server_url = 'http://' + os.getenv('VIRTUAL_HOST')
        api.app.config['REVERSE_PROXY_PATH'] = server_url
        # api.app.config['REVERSE_PROXY_PATH'] = '/api'
        ReverseProxyPrefixFix(api.app)

    # logging.info('Start spark:' + str(start_spark))
    # if start_spark:
    #     try:
    #         start_spark()
    #         logging.info('Started Spark locally')
    #     except:
    #         logging.info("Could not start Spark locally")

    print(
        "Access Swagger UI at \033[1mhttp://localhost:" + str(port) + "\033[1m 🔗")
    api.run(host='0.0.0.0', port=port, debug=debug, server=deployment_server)
    # api.run(host='0.0.0.0', port=port, debug=debug)


def post_embedding(apikey, types, emb_name, description, model_id):
    """Post JSON embeddings via the API, with simple APIKEY authentication 
    provided in environment variables 
    """
    # TODO: implement GitHub OAuth? https://github-flask.readthedocs.io/en/latest/
    # Ignore the API key check if no env variable defined (for development)
    print(os.getenv('OPENPREDICT_APIKEY'))
    if os.getenv('OPENPREDICT_APIKEY') == apikey or os.getenv('OPENPREDICT_APIKEY') is None:
        embedding_file = connexion.request.files['embedding_file']
        print(emb_name, types)
        run_id, scores = addEmbedding(
            embedding_file, emb_name, types, description, model_id)
        print('Embeddings uploaded')
        # train_model(False)
        return {
            'status': 200,
            'message': 'Embeddings added for run ' + run_id + ', trained model has scores ' + str(scores)
        }
    else:
        return {'Forbidden': 403}


def get_similarity(types='Both', drug_id=None, disease_id=None, model_id='drugs_fp_embed.txt', min_score=None, max_score=None, n_results=None):
    """Get similar entites for a given entity CURIE.

    :param entity: Search for predicted associations for this entity CURIE
    :return: Prediction results object with score
    """
    time_start = datetime.now()

    # TODO: if drug_id and disease_id defined, then check if the disease appear in the provided drug predictions
    concept_id = ''
    if drug_id:
        concept_id = drug_id
    elif disease_id:
        concept_id = disease_id
    else:
        return ('Bad request: provide a drugid or diseaseid', 400)

    try:
        emb_vectors = similarity_features[model_id]
        prediction_json, source_target_predictions = get_similarities(
            types, concept_id, emb_vectors, min_score, max_score, n_results
        )
    except Exception as e:
        print('Error processing ID ' + concept_id)
        print(e)
        return ('Not found: entry in OpenPredict for ID ' + concept_id, 404)

    # try:
    #     prediction_json = get_predictions(entity, model_id, score, n_results)
    # except:
    #     return "Not found", 404

    # relation = "biolink:treated_by"
    logging.info('PredictRuntime: ' + str(datetime.now() - time_start))
    return {'hits': prediction_json, 'count': len(prediction_json)}
    # return {'results': prediction_json, 'relation': relation, 'count': len(prediction_json)} or ('Not found', 404)




def get_predict(drug_id=None, disease_id=None, model_id='openpredict-baseline-omim-drugbank', min_score=None, max_score=None, n_results=None):
    """Get predicted associations for a given entity CURIE.

    :param entity: Search for predicted associations for this entity CURIE
    :return: Prediction results object with score
    """
    time_start = datetime.now()

    # TODO: if drug_id and disease_id defined, then check if the disease appear in the provided drug predictions
    concept_id = ''
    if drug_id:
        concept_id = drug_id
    elif disease_id:
        concept_id = disease_id
    else:
        return ('Bad request: provide a drugid or diseaseid', 400)

    try:
        prediction_json, source_target_predictions = get_predictions(
            concept_id, model_id, min_score, max_score, n_results, 
            treats_features, treats_classifier
        )
    except Exception as e:
        print('Error processing ID ' + concept_id)
        print(e)
        return ('Not found: entry in OpenPredict for ID ' + concept_id, 404)

    # try:
    #     prediction_json = get_predictions(entity, model_id, score, n_results)
    # except:
    #     return "Not found", 404

    # relation = "biolink:treated_by"
    logging.info('PredictRuntime: ' + str(datetime.now() - time_start))
    return {'hits': prediction_json, 'count': len(prediction_json)}
    # return {'results': prediction_json, 'relation': relation, 'count': len(prediction_json)} or ('Not found', 404)


def get_meta_knowledge_graph():
    """Get predicates and entities provided by the API

    :return: JSON with biolink entities
    """
    openpredict_predicates = {
        "edges": [
            {
                "object": "biolink:Disease",
                "predicate": "biolink:treats",
                "relations": [
                    "RO:0002434"
                ],
                "subject": "biolink:Drug"
            },
            {
                "object": "biolink:Drug",
                "predicate": "biolink:treated_by",
                "relations": [
                    "RO:0002434"
                ],
                "subject": "biolink:Disease"
            },
            {
                "object": "biolink:Entity",
                "predicate": "biolink:similar_to",
                # "relations": [
                #     "RO:0002434"
                # ],
                "subject": "biolink:Entity"
            },
        ],
        "nodes": {
            "biolink:Disease": {
                "id_prefixes": [
                    "OMIM"
                ]
            },
            "biolink:Drug": {
                "id_prefixes": [
                    "DRUGBANK"
                ]
            }
        }
    }

    return openpredict_predicates


def get_predicates():
    """DEPRECATED since 3.1.0, replaced by meta_knowledge_graph
    Get predicates and entities provided by the API

    :return: JSON with biolink entities
    """
    openpredict_predicates = {
        "disease": {
            "drug": [
                "treated_by"
            ]
        },
        "drug": {
            "disease": [
                "treats"
            ]
        }
    }
    return openpredict_predicates


def get_features(type):
    """Get features in the model

    :return: JSON with features
    """
    return retrieve_features(type)


def get_models():
    """Get models with their scores and features

    :return: JSON with models and features
    """
    return retrieve_models()


async def async_reasoner_predict(request_body):
    """Get predicted associations for a given ReasonerAPI query.

    :param request_body: The ReasonerStdAPI query in JSON
    :return: Predictions as a ReasonerStdAPI Message
    """
    return post_reasoner_predict(request_body)

# TODO: get_predict wrapped in ReasonerStdApi


def post_reasoner_predict(request_body):
    """Get predicted associations for a given ReasonerAPI query.

    :param request_body: The ReasonerStdAPI query in JSON
    :return: Predictions as a ReasonerStdAPI Message
    """
    query_graph = request_body["message"]["query_graph"]
    print(query_graph)
    if len(query_graph["edges"]) == 0:
        return {"message": {'knowledge_graph': {'nodes': {}, 'edges': {}}, 'query_graph': query_graph, 'results': []}}
        # return ({"status": 400, "title": "Bad Request", "detail": "No edges", "type": "about:blank" }, 400)
    if len(query_graph["edges"]) > 1:
        # Currently just return a empty result if multi-edges query
        return {"message": {'knowledge_graph': {'nodes': {}, 'edges': {}}, 'query_graph': query_graph, 'results': []}}
        # return ({"status": 501, "title": "Not Implemented", "detail": "Multi-edges queries not yet implemented", "type": "about:blank" }, 501)

    reasonerapi_response = typed_results_to_reasonerapi(request_body, treats_features, similarity_features, treats_classifier)

    # TODO: populate edges/nodes with association predictions
    #  Edge: {
    #     "id": "e50",
    #     "source_id": "MONDO:0021668",
    #     "target_id": "ChEMBL:CHEMBL560511",
    #     "type": "treated_by"
    #   }
    # Node: {
    #     "id": "ChEMBL:CHEMBL2106966",
    #     "name": "Piketoprofen",
    #     "type": "chemical_substance"
    #   },

    return reasonerapi_response or ('Not found', 404)